BMI-Calculator/requirements.txt
PyQt5
QT-PyQt-PySide-Custom-Widgets
numpy
matplotlib
mysql-connector-python
